MeshPlaneClipping processor for multiple planes, sharing clipping algorithm with existing MeshClipping processor
