+++ modules/base/CMakeLists.txt
@@ -16,6 +16,7 @@ set(HEADER_FILES
     include/modules/base/algorithm/image/layerramsubset.h
     include/modules/base/algorithm/mesh/axisalignedboundingbox.h
     include/modules/base/algorithm/mesh/meshcameraalgorithms.h
+    include/modules/base/algorithm/mesh/meshclipping.h
     include/modules/base/algorithm/mesh/meshconverter.h
     include/modules/base/algorithm/meshutils.h
     include/modules/base/algorithm/randomutils.h
@@ -73,6 +74,7 @@ set(HEADER_FILES
     include/modules/base/processors/meshexport.h
     include/modules/base/processors/meshinformation.h
     include/modules/base/processors/meshmapping.h
+    include/modules/base/processors/meshplaneclipping.h
     include/modules/base/processors/meshsequenceelementselectorprocessor.h
     include/modules/base/processors/meshsource.h
     include/modules/base/processors/noiseprocessor.h
@@ -135,6 +137,7 @@ set(SOURCE_FILES
     src/algorithm/image/layerramsubset.cpp
     src/algorithm/mesh/axisalignedboundingbox.cpp
     src/algorithm/mesh/meshcameraalgorithms.cpp
+    src/algorithm/mesh/meshclipping.cpp
     src/algorithm/mesh/meshconverter.cpp
     src/algorithm/meshutils.cpp
     src/algorithm/volume/marchingcubes.cpp
@@ -187,6 +190,7 @@ set(SOURCE_FILES
     src/processors/meshexport.cpp
     src/processors/meshinformation.cpp
     src/processors/meshmapping.cpp
+    src/processors/meshplaneclipping.cpp
     src/processors/meshsequenceelementselectorprocessor.cpp
     src/processors/meshsource.cpp
     src/processors/noiseprocessor.cpp
